Slightly changed the source tree structure
Removed the unused files and moved the display module to the root directory.
The original intention to implement a few display devices has been terminated,
as they need quite complex third-party libraries to implement the GUI. So only
the console/terminal output remained. Module may be renamed and refactored to
address this change of goal.