Removed the unused error code and fixed behavior when input file is given
At some point, I made the default dimensions of the field 10 by 10. That was
raising an error code if the input file was given, because originally I was
going to obtain the field dimensions either from the "-x -y" arguments or from
the input file. In the latter case the default values were zero.