One more cleanup commit
Fixed a race condition in the file tests. In particular, I was using the same
directory for the nested path tests, which should be deleted at the end of the
tests. That was failing when multiple threads were trying to do the same. Also
fixed a few build warnings in the given commit.